+++ python/dolfinx/io/__init__.py
@@ -9,7 +9,7 @@ from dolfinx import cpp as _cpp
 from dolfinx.cpp.io import distribute_entity_data  # noqa: F401
 from dolfinx.io.utils import VTKFile, XDMFFile  # noqa: F401
 
-__all__ = ["gmsh_io", "distribute_entity_data", "VTKFile", "XDMFFile"]
+__all__ = ["gmshio", "distribute_entity_data", "VTKFile", "XDMFFile"]
 
 if _cpp.common.has_adios2:
     # FidesWriter and VTXWriter require ADIOS2
@@ -19,5 +19,5 @@ if _cpp.common.has_adios2:
 try:
     import gmsh  # noqa: F401
 except ModuleNotFoundError:
-    from dolfinx.io import gmsh_io
-    __all__ = __all__ + ["gmsh_io"]
+    from dolfinx.io import gmshio
+    __all__ = __all__ + ["gmshio"]
