+++ python/test/unit/io/test_xdmf_mesh.py
@@ -11,7 +11,7 @@ import pytest
 
 from dolfinx import cpp as _cpp
 from dolfinx.io import XDMFFile
-from dolfinx.io.gmsh_io import cell_perm_array, ufl_mesh
+from dolfinx.io.gmshio import cell_perm_array, ufl_mesh
 from dolfinx.mesh import (CellType, GhostMode, create_mesh, create_submesh,
                           create_unit_cube, create_unit_interval,
                           create_unit_square, locate_entities)
